refactor: extract store translations
