+++ crates/bookforge-store/src/db/translations.rs
@@ -1,0 +1,691 @@
+use super::*;
+
+impl JobStore {
+    pub fn insert_segments(
+        &self,
+        job_id: &str,
+        segments: &[Segment],
+        prompt_version: &str,
+        provider: &str,
+        model: &str,
+        cache_namespace: &str,
+    ) -> Result<()> {
+        let mut conn = self.conn.borrow_mut();
+        let tx = conn.transaction()?;
+        for segment in segments {
+            tx.execute(
+                "INSERT OR IGNORE INTO segments
+                 (id, job_id, section_id, ordinal, source_hash, prompt_version, provider, model, status, attempts, cache_namespace)
+                 VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, 'queued', 0, ?9)",
+                params![
+                    segment.id.0,
+                    job_id,
+                    segment.section_id.0,
+                    segment.ordinal as i64,
+                    segment.checksum,
+                    prompt_version,
+                    provider,
+                    model,
+                    cache_namespace,
+                ],
+            )?;
+        }
+        tx.commit()?;
+        Ok(())
+    }
+
+    pub fn save_translation(&self, request: SaveTranslation<'_>) -> Result<()> {
+        if self.translation_is_human_corrected(request.job_id, request.segment_id)? {
+            return Ok(());
+        }
+        let now = timestamp_string();
+        let translated_hash = stable_hash(request.translated_text);
+        {
+            let conn = self.conn.borrow();
+            conn.execute(
+                "INSERT OR REPLACE INTO translations
+                 (segment_id, job_id, translated_text, provider, model, prompt_version, created_at)
+                 VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7)",
+                params![
+                    request.segment_id,
+                    request.job_id,
+                    request.translated_text,
+                    request.provider,
+                    request.model,
+                    request.prompt_version,
+                    now
+                ],
+            )?;
+            replace_block_translations(&conn, request.job_id, request.segment_id, request.blocks)?;
+            conn.execute(
+                "UPDATE segments
+                 SET status = 'succeeded',
+                     attempts = attempts + 1,
+                     tokens_input = ?1,
+                     tokens_input_cached = ?2,
+                     tokens_output = ?3,
+                     tokens_estimated = ?4,
+                     translated_hash = ?5,
+                     error = NULL
+                 WHERE job_id = ?6 AND id = ?7",
+                params![
+                    request.input_tokens.map(|value| value as i64),
+                    request.input_cached_tokens.map(|value| value as i64),
+                    request.output_tokens.map(|value| value as i64),
+                    if request.tokens_estimated {
+                        1_i64
+                    } else {
+                        0_i64
+                    },
+                    translated_hash,
+                    request.job_id,
+                    request.segment_id,
+                ],
+            )?;
+        }
+        self.consume_dashboard_retry_guidance(request.job_id, request.segment_id)?;
+        self.touch_job_unless_status(request.job_id, "running", &["paused", "stopped"])?;
+        Ok(())
+    }
+
+    pub fn save_needs_review(&self, request: SaveNeedsReview<'_>) -> Result<()> {
+        if self.translation_is_human_corrected(request.job_id, request.segment_id)? {
+            return Ok(());
+        }
+        let now = timestamp_string();
+        let translated_hash = stable_hash(request.preserved_text);
+        {
+            let conn = self.conn.borrow();
+            conn.execute(
+                "INSERT OR REPLACE INTO translations
+                 (segment_id, job_id, translated_text, provider, model, prompt_version, created_at)
+                 VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7)",
+                params![
+                    request.segment_id,
+                    request.job_id,
+                    request.preserved_text,
+                    request.provider,
+                    request.model,
+                    request.prompt_version,
+                    now
+                ],
+            )?;
+            replace_block_translations(&conn, request.job_id, request.segment_id, request.blocks)?;
+            conn.execute(
+                "UPDATE segments
+                 SET status = 'needs_review',
+                     attempts = attempts + 1,
+                     tokens_input = ?1,
+                     tokens_input_cached = ?2,
+                     tokens_output = ?3,
+                     tokens_estimated = ?4,
+                     translated_hash = ?5,
+                     error = ?6
+                 WHERE job_id = ?7 AND id = ?8",
+                params![
+                    request.input_tokens.map(|value| value as i64),
+                    request.input_cached_tokens.map(|value| value as i64),
+                    request.output_tokens.map(|value| value as i64),
+                    if request.tokens_estimated {
+                        1_i64
+                    } else {
+                        0_i64
+                    },
+                    translated_hash,
+                    request.error,
+                    request.job_id,
+                    request.segment_id
+                ],
+            )?;
+        }
+        self.consume_dashboard_retry_guidance(request.job_id, request.segment_id)?;
+        self.touch_job_unless_status(request.job_id, "needs_review", &["paused", "stopped"])?;
+        Ok(())
+    }
+
+    pub fn save_cached_translation(&self, request: SaveCachedTranslation<'_>) -> Result<()> {
+        if self.translation_is_human_corrected(request.job_id, request.segment_id)? {
+            return Ok(());
+        }
+        let now = timestamp_string();
+        let translated_hash = stable_hash(request.translated_text);
+        {
+            let conn = self.conn.borrow();
+            conn.execute(
+                "INSERT OR REPLACE INTO translations
+                 (segment_id, job_id, translated_text, provider, model, prompt_version, created_at)
+                 VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7)",
+                params![
+                    request.segment_id,
+                    request.job_id,
+                    request.translated_text,
+                    request.provider,
+                    request.model,
+                    request.prompt_version,
+                    now
+                ],
+            )?;
+            replace_block_translations(&conn, request.job_id, request.segment_id, request.blocks)?;
+            conn.execute(
+                "UPDATE segments
+                 SET status = 'skipped_cached',
+                     tokens_input = NULL,
+                     tokens_input_cached = NULL,
+                     tokens_output = NULL,
+                     tokens_estimated = 0,
+                     translated_hash = ?1,
+                     error = NULL
+                 WHERE job_id = ?2 AND id = ?3",
+                params![translated_hash, request.job_id, request.segment_id],
+            )?;
+        }
+        self.consume_dashboard_retry_guidance(request.job_id, request.segment_id)?;
+        self.touch_job_unless_status(request.job_id, "running", &["paused", "stopped"])?;
+        Ok(())
+    }
+
+    pub fn save_manual_correction(&self, request: SaveManualCorrection<'_>) -> Result<()> {
+        if request.translated_text.trim().is_empty() {
+            return Err(StoreError::InvalidCorrection(
+                "translation text cannot be empty".to_string(),
+            ));
+        }
+        if request.blocks.is_empty()
+            || request
+                .blocks
+                .iter()
+                .any(|block| block.text.trim().is_empty())
+        {
+            return Err(StoreError::InvalidCorrection(
+                "every corrected block must contain text".to_string(),
+            ));
+        }
+
+        let now = timestamp_string();
+        let translated_hash = stable_hash(request.translated_text);
+        let mut conn = self.conn.borrow_mut();
+        let tx = conn.transaction()?;
+        let job_status = tx
+            .query_row(
+                "SELECT status FROM jobs WHERE id = ?1",
+                params![request.job_id],
+                |row| row.get::<_, String>(0),
+            )
+            .optional()?;
+        let Some(job_status) = job_status else {
+            return Err(StoreError::InvalidCorrection(format!(
+                "job '{}' was not found",
+                request.job_id
+            )));
+        };
+        if matches!(job_status.as_str(), "running" | "paused") {
+            return Err(StoreError::InvalidCorrection(format!(
+                "job '{}' is {}; stop it before applying a manual correction",
+                request.job_id, job_status
+            )));
+        }
+
+        let prompt_version = tx
+            .query_row(
+                "SELECT prompt_version FROM segments WHERE job_id = ?1 AND id = ?2",
+                params![request.job_id, request.segment_id],
+                |row| row.get::<_, String>(0),
+            )
+            .optional()?;
+        let Some(prompt_version) = prompt_version else {
+            return Err(StoreError::InvalidCorrection(format!(
+                "segment '{}' was not found in job '{}'",
+                request.segment_id, request.job_id
+            )));
+        };
+
+        tx.execute(
+            "INSERT INTO translations
+             (segment_id, job_id, translated_text, provider, model, prompt_version, created_at,
+              origin, human_corrected, corrected_at)
+             VALUES (?1, ?2, ?3, 'manual', 'manual', ?4, ?5, 'manual', 1, ?5)
+             ON CONFLICT(job_id, segment_id) DO UPDATE SET
+               translated_text = excluded.translated_text,
+               provider = 'manual',
+               model = 'manual',
+               prompt_version = excluded.prompt_version,
+               created_at = excluded.created_at,
+               origin = 'manual',
+               human_corrected = 1,
+               corrected_at = excluded.corrected_at",
+            params![
+                request.segment_id,
+                request.job_id,
+                request.translated_text,
+                prompt_version,
+                now,
+            ],
+        )?;
+        replace_block_translations(&tx, request.job_id, request.segment_id, request.blocks)?;
+        tx.execute(
+            "UPDATE segments
+             SET status = 'succeeded', translated_hash = ?1, error = NULL
+             WHERE job_id = ?2 AND id = ?3",
+            params![translated_hash, request.job_id, request.segment_id],
+        )?;
+        tx.execute(
+            "DELETE FROM qa_findings WHERE job_id = ?1 AND segment_id = ?2",
+            params![request.job_id, request.segment_id],
+        )?;
+        tx.execute(
+            "UPDATE segment_flags SET consumed = 1 WHERE job_id = ?1 AND segment_id = ?2",
+            params![request.job_id, request.segment_id],
+        )?;
+        tx.commit()?;
+        drop(conn);
+
+        self.recompute_job_status(request.job_id)?;
+        Ok(())
+    }
+
+    pub fn translation_is_human_corrected(&self, job_id: &str, segment_id: &str) -> Result<bool> {
+        let conn = self.conn.borrow();
+        Ok(conn
+            .query_row(
+                "SELECT human_corrected FROM translations WHERE job_id = ?1 AND segment_id = ?2",
+                params![job_id, segment_id],
+                |row| row.get::<_, i64>(0),
+            )
+            .optional()?
+            .is_some_and(|value| value != 0))
+    }
+
+    fn consume_dashboard_retry_guidance(&self, job_id: &str, segment_id: &str) -> Result<()> {
+        let conn = self.conn.borrow();
+        conn.execute(
+            "UPDATE segment_flags SET consumed = 1
+             WHERE job_id = ?1 AND segment_id = ?2 AND kind = 'dashboard_retry'",
+            params![job_id, segment_id],
+        )?;
+        Ok(())
+    }
+
+    pub fn pending_segment_ids(&self, job_id: &str) -> Result<Vec<String>> {
+        let conn = self.conn.borrow();
+        let mut stmt = conn.prepare(
+            "SELECT id FROM segments
+             WHERE job_id = ?1 AND status IN ('queued', 'retry_pending')
+             ORDER BY ordinal",
+        )?;
+        let rows = stmt.query_map(params![job_id], |row| row.get::<_, String>(0))?;
+        rows.collect::<std::result::Result<Vec<_>, _>>()
+            .map_err(StoreError::from)
+    }
+
+    pub fn segment_records(&self, job_id: &str) -> Result<Vec<SegmentRecord>> {
+        let conn = self.conn.borrow();
+        let mut stmt = conn.prepare(
+            "SELECT id,
+                    status,
+                    attempts,
+                    error,
+                    COALESCE(tokens_input, input_tokens),
+                    tokens_input_cached,
+                    COALESCE(tokens_output, output_tokens),
+                    tokens_estimated
+             FROM segments WHERE job_id = ?1 ORDER BY ordinal",
+        )?;
+        let rows = stmt.query_map(params![job_id], |row| {
+            Ok(SegmentRecord {
+                id: row.get(0)?,
+                status: row.get(1)?,
+                attempts: row.get::<_, i64>(2)? as usize,
+                error: row.get(3)?,
+                input_tokens: row.get::<_, Option<i64>>(4)?.map(|value| value as u64),
+                input_cached_tokens: row.get::<_, Option<i64>>(5)?.map(|value| value as u64),
+                output_tokens: row.get::<_, Option<i64>>(6)?.map(|value| value as u64),
+                tokens_estimated: row.get::<_, i64>(7)? != 0,
+            })
+        })?;
+        rows.collect::<std::result::Result<Vec<_>, _>>()
+            .map_err(StoreError::from)
+    }
+
+    pub fn load_block_translations(&self, job_id: &str) -> Result<Vec<StoredBlockTranslation>> {
+        let conn = self.conn.borrow();
+        let mut stmt = conn.prepare(
+            "SELECT segment_id, block_id, translated_text
+             FROM translation_blocks WHERE job_id = ?1 ORDER BY segment_id, block_id",
+        )?;
+        let rows = stmt.query_map(params![job_id], |row| {
+            Ok(StoredBlockTranslation {
+                segment_id: row.get(0)?,
+                block_id: row.get(1)?,
+                text: row.get(2)?,
+            })
+        })?;
+        rows.collect::<std::result::Result<Vec<_>, _>>()
+            .map_err(StoreError::from)
+    }
+
+    pub fn load_terminal_segment_translations(
+        &self,
+        job_id: &str,
+    ) -> Result<Vec<StoredSegmentTranslation>> {
+        let conn = self.conn.borrow();
+        let mut stmt = conn.prepare(
+            "SELECT s.id, s.ordinal, s.status, s.error, t.translated_text,
+                    t.provider, t.model, t.human_corrected, t.corrected_at
+             FROM segments s
+             JOIN translations t ON t.job_id = s.job_id AND t.segment_id = s.id
+             WHERE s.job_id = ?1 AND s.status IN ('succeeded', 'skipped_cached', 'needs_review')
+             ORDER BY s.ordinal",
+        )?;
+        let rows = stmt.query_map(params![job_id], |row| {
+            Ok((
+                row.get::<_, String>(0)?,
+                row.get::<_, i64>(1)?,
+                row.get::<_, String>(2)?,
+                row.get::<_, Option<String>>(3)?,
+                row.get::<_, String>(4)?,
+                row.get::<_, String>(5)?,
+                row.get::<_, String>(6)?,
+                row.get::<_, i64>(7)? != 0,
+                row.get::<_, Option<String>>(8)?,
+            ))
+        })?;
+
+        let mut records = Vec::new();
+        for row in rows {
+            let (
+                segment_id,
+                ordinal,
+                status,
+                error,
+                translated_text,
+                provider,
+                model,
+                human_corrected,
+                corrected_at,
+            ) = row?;
+            let mut block_stmt = conn.prepare(
+                "SELECT block_id, translated_text
+                 FROM translation_blocks
+                 WHERE job_id = ?1 AND segment_id = ?2
+                 ORDER BY block_id",
+            )?;
+            let blocks = block_stmt
+                .query_map(params![job_id, segment_id.as_str()], |row| {
+                    Ok(BlockTranslation {
+                        block_id: BlockId(row.get::<_, String>(0)?),
+                        text: row.get(1)?,
+                    })
+                })?
+                .collect::<std::result::Result<Vec<_>, _>>()?;
+            records.push(StoredSegmentTranslation {
+                segment_id,
+                ordinal: ordinal as usize,
+                status,
+                error,
+                translated_text,
+                blocks,
+                provider,
+                model,
+                human_corrected,
+                corrected_at,
+            });
+        }
+
+        Ok(records)
+    }
+
+    pub fn resumable_segment_ids(&self, job_id: &str) -> Result<Vec<String>> {
+        let conn = self.conn.borrow();
+        let mut stmt = conn.prepare(
+            "SELECT id FROM segments
+             WHERE job_id = ?1 AND status IN ('queued', 'retry_pending', 'failed')
+             ORDER BY ordinal",
+        )?;
+        let rows = stmt.query_map(params![job_id], |row| row.get::<_, String>(0))?;
+        rows.collect::<std::result::Result<Vec<_>, _>>()
+            .map_err(StoreError::from)
+    }
+
+    #[allow(clippy::too_many_arguments)]
+    pub fn find_cached_translation(
+        &self,
+        segment: &Segment,
+        prompt_version: &str,
+        provider: &str,
+        model: &str,
+        source_lang: Option<&str>,
+        target_lang: &str,
+        cache_namespace: &str,
+    ) -> Result<Option<CachedTranslation>> {
+        let conn = self.conn.borrow();
+        let cached = conn
+            .query_row(
+                "SELECT t.job_id, t.segment_id, t.translated_text
+                 FROM translations t
+                 JOIN segments s ON s.job_id = t.job_id AND s.id = t.segment_id
+                 JOIN jobs j ON j.id = t.job_id
+                 WHERE s.source_hash = ?1
+                   AND s.prompt_version = ?2
+                   AND s.provider = ?3
+                   AND s.model = ?4
+                   AND ((?5 IS NULL AND j.source_lang IS NULL) OR j.source_lang = ?5)
+                   AND j.target_lang = ?6
+                   AND s.cache_namespace = ?7
+                   AND s.status IN ('succeeded', 'skipped_cached')
+                   AND t.human_corrected = 0
+                 ORDER BY CASE s.status WHEN 'succeeded' THEN 0 ELSE 1 END,
+                          CAST(t.created_at AS INTEGER) DESC,
+                          t.rowid DESC
+                 LIMIT 1",
+                params![
+                    segment.checksum,
+                    prompt_version,
+                    provider,
+                    model,
+                    source_lang,
+                    target_lang,
+                    cache_namespace,
+                ],
+                |row| {
+                    Ok((
+                        row.get::<_, String>(0)?,
+                        row.get::<_, String>(1)?,
+                        row.get::<_, String>(2)?,
+                    ))
+                },
+            )
+            .optional()?;
+
+        let Some((job_id, segment_id, translated_text)) = cached else {
+            return Ok(None);
+        };
+
+        let mut stmt = conn.prepare(
+            "SELECT block_id, translated_text
+             FROM translation_blocks
+             WHERE job_id = ?1 AND segment_id = ?2
+             ORDER BY block_id",
+        )?;
+        let rows = stmt.query_map(params![job_id, segment_id], |row| {
+            Ok(BlockTranslation {
+                block_id: BlockId(row.get::<_, String>(0)?),
+                text: row.get(1)?,
+            })
+        })?;
+        let blocks = rows.collect::<std::result::Result<Vec<_>, _>>()?;
+
+        let mut by_id = blocks
+            .into_iter()
+            .map(|block| (block.block_id.0.clone(), block))
+            .collect::<HashMap<_, _>>();
+
+        let mut ordered = Vec::with_capacity(segment.block_ids.len());
+        for id in &segment.block_ids {
+            let Some(block) = by_id.remove(&id.0) else {
+                return Ok(None);
+            };
+            ordered.push(block);
+        }
+        if !by_id.is_empty() {
+            return Ok(None);
+        }
+
+        Ok(Some(CachedTranslation {
+            translated_text,
+            blocks: ordered,
+        }))
+    }
+
+    pub fn find_cached_translations_batch(
+        &self,
+        segments: &[Segment],
+        request: CacheLookupRequest<'_>,
+    ) -> Result<HashMap<String, CachedTranslation>> {
+        let mut results = HashMap::new();
+        if segments.is_empty() {
+            return Ok(results);
+        }
+
+        const SQLITE_IN_CHUNK_SIZE: usize = 900;
+
+        for chunk in segments.chunks(SQLITE_IN_CHUNK_SIZE) {
+            let hashes: Vec<&str> = chunk.iter().map(|s| s.checksum.as_str()).collect();
+            let placeholders: Vec<String> =
+                (0..hashes.len()).map(|i| format!("?{}", i + 1)).collect();
+            let placeholders_sql = placeholders.join(", ");
+
+            let sql = format!(
+                "SELECT t.job_id, t.segment_id, t.translated_text, s.source_hash
+                 FROM translations t
+                 JOIN segments s ON s.job_id = t.job_id AND s.id = t.segment_id
+                 JOIN jobs j ON j.id = t.job_id
+                 WHERE s.source_hash IN ({placeholders_sql})
+                   AND s.prompt_version = ?{}
+                   AND s.provider = ?{}
+                   AND s.model = ?{}
+                   AND ((?{} IS NULL AND j.source_lang IS NULL) OR j.source_lang = ?{})
+                   AND j.target_lang = ?{}
+                   AND s.cache_namespace = ?{}
+                   AND s.status IN ('succeeded', 'skipped_cached')
+                   AND t.human_corrected = 0
+                 ORDER BY CASE s.status WHEN 'succeeded' THEN 0 ELSE 1 END,
+                          CAST(t.created_at AS INTEGER) DESC,
+                          t.rowid DESC",
+                hashes.len() + 1,
+                hashes.len() + 2,
+                hashes.len() + 3,
+                hashes.len() + 4,
+                hashes.len() + 5,
+                hashes.len() + 6,
+                hashes.len() + 7,
+            );
+
+            let mut params: Vec<Box<dyn rusqlite::types::ToSql>> = Vec::new();
+            for hash in &hashes {
+                params.push(Box::new(hash.to_string()));
+            }
+            params.push(Box::new(request.prompt_version.to_string()));
+            params.push(Box::new(request.provider.to_string()));
+            params.push(Box::new(request.model.to_string()));
+            params.push(Box::new(request.source_lang.map(|s| s.to_string())));
+            params.push(Box::new(request.source_lang.map(|s| s.to_string())));
+            params.push(Box::new(request.target_lang.to_string()));
+            params.push(Box::new(request.cache_namespace.to_string()));
+
+            let conn = self.conn.borrow();
+            let mut stmt = conn.prepare(&sql)?;
+            let param_refs: Vec<&dyn rusqlite::types::ToSql> =
+                params.iter().map(|p| p.as_ref()).collect();
+
+            let rows = stmt.query_map(param_refs.as_slice(), |row| {
+                Ok((
+                    row.get::<_, String>(0)?,
+                    row.get::<_, String>(1)?,
+                    row.get::<_, String>(2)?,
+                    row.get::<_, String>(3)?,
+                ))
+            })?;
+
+            let mut hash_to_hit: HashMap<String, (String, String, String)> = HashMap::new();
+            for row in rows {
+                let (job_id, segment_id, translated_text, source_hash) = row?;
+                hash_to_hit
+                    .entry(source_hash)
+                    .or_insert((job_id, segment_id, translated_text));
+            }
+
+            for segment in chunk {
+                if let Some((job_id, segment_id, translated_text)) =
+                    hash_to_hit.get(&segment.checksum)
+                {
+                    let mut block_stmt = conn.prepare(
+                        "SELECT block_id, translated_text
+                         FROM translation_blocks
+                         WHERE job_id = ?1 AND segment_id = ?2
+                         ORDER BY block_id",
+                    )?;
+                    let block_rows = block_stmt.query_map(params![job_id, segment_id], |row| {
+                        Ok(BlockTranslation {
+                            block_id: BlockId(row.get::<_, String>(0)?),
+                            text: row.get(1)?,
+                        })
+                    })?;
+                    let blocks = block_rows.collect::<std::result::Result<Vec<_>, _>>()?;
+
+                    let mut by_id = blocks
+                        .into_iter()
+                        .map(|block| (block.block_id.0.clone(), block))
+                        .collect::<HashMap<_, _>>();
+
+                    let mut ordered = Vec::with_capacity(segment.block_ids.len());
+                    let mut valid = true;
+                    for id in &segment.block_ids {
+                        let Some(block) = by_id.remove(&id.0) else {
+                            valid = false;
+                            break;
+                        };
+                        ordered.push(block);
+                    }
+                    if !valid || !by_id.is_empty() {
+                        continue;
+                    }
+
+                    results.insert(
+                        segment.id.0.clone(),
+                        CachedTranslation {
+                            translated_text: translated_text.clone(),
+                            blocks: ordered,
+                        },
+                    );
+                }
+            }
+        }
+
+        Ok(results)
+    }
+}
+
+fn replace_block_translations(
+    conn: &Connection,
+    job_id: &str,
+    segment_id: &str,
+    blocks: &[BlockTranslation],
+) -> rusqlite::Result<()> {
+    conn.execute(
+        "DELETE FROM translation_blocks WHERE job_id = ?1 AND segment_id = ?2",
+        params![job_id, segment_id],
+    )?;
+    for block in blocks {
+        conn.execute(
+            "INSERT INTO translation_blocks (segment_id, job_id, block_id, translated_text)
+             VALUES (?1, ?2, ?3, ?4)",
+            params![
+                segment_id,
+                job_id,
+                block.block_id.0.as_str(),
+                block.text.as_str()
+            ],
+        )?;
+    }
+    Ok(())
+}
